Clean up all compiler warnings in our own code
Reviewed every warning from a full clean build against the now-pinned
SDK and fixed the ones in code we own (left vendored lib/ warnings
alone - not worth diverging further from upstream for cosmetics):

- CMakeLists.txt: add_custom_command(TARGET ...) doesn't accept
  DEPENDS and was missing an explicit POST_BUILD, both silently
  papered over by CMake for backward compatibility (CMP0175 warning).
  The APP_HDR dependency is already established via the preceding
  add_custom_target().
- SettingsFlash.cpp: explicit static_cast<uint8_t> on the address
  bytes split into the SPI command buffers - the truncation was
  always intentional (that's how a 24-bit flash address gets split
  into 3 bytes), just not spelled out for -Wnarrowing.
- SettingsManager.cpp: mark the device-id variable [[maybe_unused]]
  - it's only read by USB_PRINTF, which compiles away entirely
    without USB_DEBUG defined.
- smart_lcc.cpp: this one wasn't just cosmetic. The vendored
  no-OS-FatFS version here added required fields to sd_sdio_if_t/
  sd_card_t (SDIO_PIO, DMA_IRQ_num, drive strengths, card-detect
  wiring) that our initializer never set. Left at their implicit
  zero-default, CLK_gpio/D1_gpio/D2_gpio/D3_gpio would all have been
  GPIO 0 (which is ESP_TX) and SDIO_PIO a null PIO pointer - the
  driver now reads these directly instead of deriving them from
  D0_gpio itself. This was a real latent bug, just dormant because
  SD_DET_B currently reports no card inserted so the code path never
  runs. Filled in the pin fields using pins.h's existing named
  SD_SCLK/SD_DAT1/SD_DAT2/SD_CS_DAT3 constants (which already match
  the documented D0-relative offsets), pio0/DMA_IRQ_0 (unused
  elsewhere in this project), set_drive_strength=false (no
  board-specific tuning known, so use RP2040 reset defaults rather
  than guess a value), and use_card_detect=false (unchanged behavior
  - sd_card_inserted() already gates mounting manually in main1()).
  Every remaining field is explicitly documented in sd_card.h as
  driver-owned state, zero/null-initialized same as before.

Deliberately NOT touched: the 5 unhandled-enum warnings in
EspFirmware.cpp for ESP_MESSAGE_ADD_COMMAND_TO_ROUTINE_STEP,
ESP_MESSAGE_ADD_EXIT_CONDITION_TO_ROUTINE_STEP, ESP_MESSAGE_POLL_STATUS,
ESP_MESSAGE_NOT_USED and ESP_SYSTEM_COMMAND_CANCEL_ROUTINE - these mark
exactly the routine-engine wiring planned for the next phase of work,
so a throwaway no-op case now would just be redone properly there.
Also left SystemController.cpp's unused 'minutes' parameter in
setAutoSleepMinutes(), part of the already-identified dead
auto-sleep code path that's intentionally parked, not implemented as
a side effect of a warning cleanup pass.

No behavior changes beyond the smart_lcc.cpp latent-bug fix (which
only matters once an SD card is actually used, and even then now
uses the same-or-safer inert defaults on every field it wasn't
already using).

+++ src/Controller/Core1/SettingsFlash.cpp
@@ -55,9 +55,9 @@ void SettingsFlash::read(uint32_t addr, uint8_t *buf, size_t len) {
     cs_select(_csPin);
     uint8_t cmdbuf[4] = {
             FLASH_CMD_READ,
-            addr >> 16,
-            addr >> 8,
-            addr
+            static_cast<uint8_t>(addr >> 16),
+            static_cast<uint8_t>(addr >> 8),
+            static_cast<uint8_t>(addr)
     };
     spi_write_blocking(_spi, cmdbuf, 4);
     spi_read_blocking(_spi, 0, buf, len);
@@ -85,9 +85,9 @@ void SettingsFlash::wait_done() {
 void SettingsFlash::sector_erase(uint32_t addr) {
     uint8_t cmdbuf[4] = {
             FLASH_CMD_SECTOR_ERASE,
-            addr >> 16,
-            addr >> 8,
-            addr
+            static_cast<uint8_t>(addr >> 16),
+            static_cast<uint8_t>(addr >> 8),
+            static_cast<uint8_t>(addr)
     };
     write_enable();
     cs_select(_csPin);
@@ -105,9 +105,9 @@ void SettingsFlash::page_program(uint32_t addr, uint8_t *buf, size_t len) {
 
     uint8_t cmdbuf[4] = {
             FLASH_CMD_PAGE_PROGRAM,
-            addr >> 16,
-            addr >> 8,
-            addr
+            static_cast<uint8_t>(addr >> 16),
+            static_cast<uint8_t>(addr >> 8),
+            static_cast<uint8_t>(addr)
     };
     write_enable();
     wait_done();
